Module bevy_inspector_egui::bevy_inspector

Methods for displaying bevy resources, assets and entities

Example

use bevy_inspector_egui::bevy_inspector;

#[derive(Debug, Clone, Eq, PartialEq, Hash, Reflect)]
enum AppState { A, B, C }

fn show_ui(world: &mut World, ui: &mut egui::Ui) {
    let mut any_reflect_value = Vec3::new(1.0, 2.0, 3.0);
    bevy_inspector::ui_for_value(&mut any_reflect_value, ui, world);

    ui.heading("Msaa resource");
    bevy_inspector::ui_for_resource::<Msaa>(world, ui);

    ui.heading("App State");
    bevy_inspector::ui_for_state::<AppState>(world, ui);

    egui::CollapsingHeader::new("Entities")
        .default_open(true)
        .show(ui, |ui| {
            bevy_inspector::ui_for_world_entities(world, ui);
        });
    egui::CollapsingHeader::new("Resources").show(ui, |ui| {
        bevy_inspector::ui_for_resources(world, ui);
    });
    egui::CollapsingHeader::new("Assets").show(ui, |ui| {
        bevy_inspector::ui_for_all_assets(world, ui);
    });
}

Modules

by_type_id

hierarchy

UI for displaying the entity hierarchy

short_circuit

Short circuiting methods for the InspectorUi to enable it to display Handles

Functions

guess_entity_name

Guesses an appropriate entity name like Light (6) or falls back to Entity (8)

ui_for_all_assets

Display all reflectable assets

ui_for_assets

Display all assets of the specified asset type A

ui_for_entities_shared_components

Display the given entity with all its components and children

ui_for_entity

Display the components of the given entity

ui_for_entity_with_children

Display the given entity with all its components and children

ui_for_resource

Display the resource R

ui_for_resources

Display all reflectable resources in the world

ui_for_state

Display state T and change state on edit

ui_for_value

Display a single &mut dyn Reflect.

ui_for_world

Display Entities, Resources and Assets using their respective functions inside headers

ui_for_world_entities

Display all entities and their components

ui_for_world_entities_filtered

Display all entities matching the given filter