bevy_inspector_egui::bevy_inspector
Function ui_for_value

pub fn ui_for_value(
    value: &mut dyn Reflect,
    ui: &mut Ui,
    world: &mut World,
) -> bool

Display a single &mut dyn Reflect.

If you are wondering why this function takes in a &mut World, it’s so that if the value contains e.g. a Handle<StandardMaterial> it can look up the corresponding asset resource and display the asset value inline.

If all you’re displaying is a simple value without any references into the bevy world, consider just using reflect_inspector::ui_for_value.