bevy_inspector_egui
Module quick

Easy plugins for showing UI panels.

Pros: no manual code required

Cons: not configurable

When you want something more custom, you can use these plugins as a starting point.

Structs

AssetInspectorPlugin

Plugin displaying an egui window for all assets of type A. Remember to call App::register_asset_reflect.

FilterQueryInspectorPlugin

Plugin displaying an egui window for all entities matching the filter F.

ResourceInspectorPlugin

Plugin displaying an egui window for a single resource. Remember to insert the resource and call App::register_type.

StateInspectorPlugin

Plugin displaying an egui window for an app state. Remember to call App::add_state.

WorldInspectorPlugin

Plugin displaying a egui window with an entity list, resources and assets