Function bevy_inspector_egui::reflect_inspector::ui_for_value

pub fn ui_for_value(
    value: &mut dyn Reflect,
    ui: &mut Ui,
    type_registry: &TypeRegistry
) -> bool

Display the value without any Context or short circuiting behaviour. This means that for example bevy’s Handle<StandardMaterial> values cannot be displayed, as they would need to have access to the World.

Use InspectorUi::new instead to provide context or use one of the methods in bevy_inspector.