bevy_inspector_egui::reflect_inspector
Function ui_for_value_readonly

pub fn ui_for_value_readonly(
    value: &dyn Reflect,
    ui: &mut Ui,
    type_registry: &TypeRegistry,
)

Display the readonly value without any Context or short circuiting behaviour. This means that for example bevy’s Handle<StandardMaterial> values cannot be displayed, as they would need to have access to the World.

Use InspectorUi::new instead to provide context or use one of the methods in bevy_inspector.