Module bevy_inspector_egui::inspector_egui_impls

Custom UI implementations for specific types. Check InspectorPrimitive for an example.

Structs

InspectorEguiImpl

Function pointers for displaying a concrete type, to be registered in the TypeRegistry.

Traits

InspectorPrimitive

Custom UI implementation for a concrete type.

Functions

register_bevy_impls

Register InspectorEguiImpls for bevy types

register_glam_impls

Register InspectorEguiImpls for bevy_math/glam types

register_std_impls

Register InspectorEguiImpls for primitive rust types as well as standard library types